add ability to move up/down between response/prompt window, reserve tab to toggle between normal/insert mode

+++ lumni/src/apps/builtin/llm/prompt/src/tui/events/text_window_event.rs
@@ -142,12 +142,12 @@ where
         }
         'j' => {
             let lines_to_move =
-                key_track.take_numeric_input().unwrap_or(1) as u16;
+                key_track.take_numeric_input().unwrap_or(1);
             window.move_cursor(MoveCursor::Down(lines_to_move));
         }
         'k' => {
             let lines_to_move =
-                key_track.take_numeric_input().unwrap_or(1) as u16;
+                key_track.take_numeric_input().unwrap_or(1);
             window.move_cursor(MoveCursor::Up(lines_to_move));
         }
         'v' => {
